requirements.txt
certifi
chardet
charset-normalizer
coloredlogs
contourpy
cycler
flatbuffers
fonttools
humanfriendly
idna
joblib
jsonpickle
kiwisolver
matplotlib
more-itertools
mpmath
music21
numpy>=1.24.0
scipy>=1.10.0
oemer
opencv-python-headless
packaging
pillow
protobuf
pygame
pyparsing
pyreadline3
python-dateutil
requests
scikit-learn
scipy
setuptools
six
sympy
threadpoolctl
types-Pillow
types-protobuf
types-requests
types-tensorflow
typing-extensions
urllib3
webcolors
opencv-python
onnxruntime
fastapi
uvicorn
python-multipart